Backend/requirements.txt
python-dotenv
langchain_openai
langchain
langgraph
pyprojroot
langchain_community
langchain_chroma
websockets